# core/api/docs/v2/openapi.yaml
components:
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: Invalid path, query or request input
    Unauthorized:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
      description: Missing, unknown, revoked or expired key; shared authentication envelope
...
  schemas:
...
    SchemaIndex:
      properties:
        kinds:
          items:
            $ref: '#/components/schemas/SchemaIndexEntry'
          type: array
          uniqueItems: false
        ops:
          items:
            $ref: '#/components/schemas/SchemaIndexEntry'
          type: array
          uniqueItems: false
      type: object
...
  securitySchemes:
    bearerauth:
      scheme: bearer
      type: http
...
info:
  contact:
    email: support@anytype.io
    name: Anytype Support
    url: https://anytype.io/contact
  description: |-
    The agent-oriented Anytype local API. An object is one AnyBlock JSON document rather than a tree of blocks, so a single GET returns a whole editable document and a single PATCH edits it.
    Everything this API names for itself is snake_case: path and query parameters, request and response fields, and the names of the PATCH ops. Things are addressed by name rather than by id: property keys, option names, and `type` as a type key. Object ids remain ids.
    Property and type keys have two spellings, and a served body picks one. The default is the key: a snake_case slug minted when the property was created and frozen from then on, so it survives every rename. Store that spelling in a config, a script, an integration. `?keys=name` on a read serves the display-name vocabulary instead, "Due date" rather than due_date, which is the spelling to show a user their own words. Input needs no such choice: every channel accepts both, plus the forgiving fold between them, so dueDate, due-date, "Due date" and Дата выполнения all address the property they name. A spelling matching two properties is a 400 listing both, and an unknown one never creates anything.
...
    A read never fails on content it cannot represent. Whatever a representation cannot express is reported in `warnings` beside the result.
...
  license:
    name: Any Source Available License 1.0
...
openapi: 3.1.0
paths:
...
  /v2/schemas:
    get:
      operationId: list_schemas
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaIndex'
          description: Schema index
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
      security:
      - bearerauth: []
      summary: List the available schemas
      tags:
      - Schemas
  /v2/schemas/{kind}:
    get:
      operationId: get_schema
      parameters:
      - description: Schema kind, as listed by GET /v2/schemas
        in: path
        name: kind
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaEntry'
          description: Schema + example
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unknown kind
      security:
      - bearerauth: []
      summary: Get the schema for one kind
      tags:
      - Schemas
  /v2/schemas/ops/{op}:
    get:
      description: The example is a single op object, ready to drop into an edit request's
        `ops` array, not a whole request body.
      operationId: get_op_schema
      parameters:
      - description: 'Op name: set_properties, update_block, replace_subtree, insert_blocks,
          move_block, delete_block, replace_text, set_cell, update_view, insert_view,
          move_view, delete_view, add_items, remove_items'
        in: path
        name: op
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaEntry'
          description: Schema + example
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unknown op
      security:
      - bearerauth: []
      summary: Get the schema for one edit op
      tags:
      - Schemas
...
  /v2/spaces/{space_id}/collections:
    post:
      description: Item ids are checked against the space; an id that does not resolve
        there is refused rather than dropped.
      operationId: create_collection
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      - description: 'Create select options for names the property does not hold yet
          (default false: an unmatched name is refused)'
        in: query
        name: create_missing_options
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Dry run; validation result without committing
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Created collection id
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation or reference failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "413":
          $ref: '#/components/responses/RequestTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Create a collection
      tags:
      - Lists
...
  /v2/spaces/{space_id}/files:
    post:
      description: 'Send multipart/form-data with a `file` field, or JSON {"url":
        …}. A source that refuses the fetch, or a URL that cannot be fetched, is a
        400 naming /url; only a genuine server fault answers 500. The id that comes
        back is the one file blocks, image blocks and icon_image values reference.'
      operationId: upload_file
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                name:
                  type: string
                url:
                  type: string
              required:
              - url
              type: object
          multipart/form-data:
            schema:
              additionalProperties: false
              properties:
                file:
                  format: binary
                  type: string
              required:
              - file
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResult'
          description: Dry run; validation result without committing
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResult'
          description: Created file object id
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation failure, or a source URL that did not yield the
            file
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "413":
          $ref: '#/components/responses/RequestTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Upload a file
      tags:
      - Files
...
  /v2/spaces/{space_id}/objects:
    get:
      operationId: list_objects
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Comma-separated property keys to include per row
        in: query
        name: fields
        schema:
          type: string
      - description: Items to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Items to return
        in: query
        name: limit
        schema:
          default: 25
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse-ObjectRow'
          description: Minimal object rows
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid fields query
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Space not found or unavailable
      security:
      - bearerauth: []
      summary: List the objects in a space
      tags:
      - Objects
    post:
      description: A select value naming an option the property does not hold is refused
        unless `create_missing_options=true` is set. An unknown type or property key
        is rejected either way, with the closest matches named. The body is either
        a full AnyBlock document or the shortcut {type, name, properties, markdown};
        `formatVersion` or `blocks` picks the document form.
      operationId: create_object
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      - description: 'Create select options for names the property does not hold yet
          (default false: an unmatched name is refused)'
        in: query
        name: create_missing_options
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Dry run; validation result without committing
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Created object id + etag
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation or reference failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Create an object
      tags:
      - Objects
...
  /v2/spaces/{space_id}/properties:
    get:
      operationId: list_properties
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse-PropertyRow'
          description: Property rows
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Space not found or unavailable
      security:
      - bearerauth: []
      summary: List the properties in a space
      tags:
      - Properties
    post:
      operationId: create_property
      parameters:
      - description: Space id
        in: path
        name: space_id
        required: true
        schema:
          type: string
      - description: Validate and report without committing
        in: query
        name: dry_run
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Dry run; validation result without committing
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResult'
          description: Created property id + key
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Validation failure
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "413":
          $ref: '#/components/responses/RequestTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
      security:
      - bearerauth: []
      summary: Create a property
      tags:
      - Properties
...
servers:
...